+++ riko/_io.py
@@ -267,7 +267,7 @@ def opener(  # noqa: E302
     r = None
 
     if url.startswith("http"):
-        r = requests.get(url, params=params, stream=binary, timeout=timeout)
+        r = requests.get(url, params=params, stream=not memoize, timeout=timeout)
         r.raise_for_status()
         r.raw.decode_content = True